+++ workers/contributor_worker/setup.py
@@ -29,7 +29,7 @@ setup(
         'requests==2.22.0',
         'psycopg2-binary==2.8.6',
         'click==7.1.2', 
-        'scipy==1.7.0',
+        'scipy==1.4.1',
         'sklearn==0.0'
     ],
     entry_points={
