--- workers/deps_worker/deps_worker-model2.py
@@ -1,141 +1,0 @@
-#SPDX-License-Identifier: MIT
-import os, subprocess
-from datetime import datetime
-import logging
-import requests
-import json
-from urllib.parse import quote
-from multiprocessing import Process, Queue
-
-import pandas as pd
-import sqlalchemy as s
-from sqlalchemy.ext.automap import automap_base
-from sqlalchemy import MetaData
-from workers.worker_base import Worker
-
-from workers.deps_worker import dependancy_calculator as dep_calc
-
-class DepsWorker(Worker):
-    def __init__(self, config={}):
-
-        worker_type = "deps_worker"
-
-        # Define what this worker can be given and know how to interpret
-        given = [['git_url']]
-        #models = ['deps']
-        models = ['deps', 'ossf_scorecard']
-
-        # Define the tables needed to insert, update, or delete on
-        #data_tables = ['repo_dependencies'] #TODO -- Add any new tables you put data into
-        data_tables = ['repo_dependencies', 'repo_deps_scorecard'] 
-        operations_tables = ['worker_history', 'worker_job']
-
-
-        # Run the general worker initialization
-        super().__init__(worker_type, config, given, models, data_tables, operations_tables)
-
-        self.config.update({
-            'repo_directory': self.augur_config.get_value('Workers', 'facade_worker')['repo_directory']
-        })
-
-        self.tool_source = 'Deps Worker'
-        self.tool_version = '2.0.0'
-        self.data_source = 'Augur Repository Data'
-
-    def deps_model(self, entry_info, repo_id):
-        """ Data collection and storage method
-        """
-        self.logger.info(entry_info)
-        self.logger.info(repo_id)
-
-        repo_path_sql = s.sql.text("""
-            SELECT repo_id, CONCAT(repo_group_id || chr(47) || repo_path || repo_name) AS path
-            FROM repo
-            WHERE repo_id = :repo_id
-        """)
-
-        relative_repo_path = self.db.execute(repo_path_sql, {'repo_id': repo_id}).fetchone()[1]
-        absolute_repo_path = self.config['repo_directory'] + relative_repo_path
-
-        try:
-            self.generate_deps_data(repo_id, absolute_repo_path)
-        except Exception as e:
-            self.logger.error(e)
-
-        self.register_task_completion(entry_info, repo_id, "deps")
-
-    def ossf_scorecard_model(self, entry_info, repo_id):
-        """ Data collection and storage method
-        """
-        self.logger.info(entry_info)
-        self.logger.info(repo_id)
-
-        repo_path_sql = s.sql.text("""
-            SELECT repo_id, CONCAT(repo_group_id || chr(47) || repo_path || repo_name) AS path
-            FROM repo
-            WHERE repo_id = :repo_id
-        """)
-
-        relative_repo_path = self.db.execute(repo_path_sql, {'repo_id': repo_id}).fetchone()[1]
-        absolute_repo_path = self.config['repo_directory'] + relative_repo_path
-
-## TODO: Flesh out generate_scorecard 
-## You can look at the Value worker to see how Go Programs are already called in Augur. 
-        try:
-            self.generate_scorecard(repo_id, absolute_repo_path)
-        except Exception as e:
-            self.logger.error(e)
-
-        self.register_task_completion(entry_info, repo_id, "deps")
-
-    def generate_scorecard(self, repo_id, path): 
-        """Runs scc on repo and stores data in database
-
-        :param repo_id: Repository ID
-        :param path: Absolute path of the Repostiory
-        """
-        self.logger.info('Searching for deps in repo')
-        self.logger.info(f'Repo ID: {repo_id}, Path: {path}')
-
-
-        # for dep in deps:
-        #         repo_deps = {
-        #             'repo_id': repo_id,
-        #             'dep_name' : dep.name,
-        #             'dep_count' : dep.count,
-        #             'dep_language' : dep.language,
-        #             'tool_source': self.tool_source,
-        #             'tool_version': self.tool_version,
-        #             'data_source': self.data_source,
-        #             'data_collection_date': datetime.now().strftime('%Y-%m-%dT%H:%M:%SZ')
-        #         }
-
-        #         result = self.db.execute(self.repo_dependencies_table.insert().values(repo_deps))
-        #         self.logger.info(f"Added dep: {result.inserted_primary_key}")
-
-
-    def generate_deps_data(self, repo_id, path):
-        """Runs scc on repo and stores data in database
-
-        :param repo_id: Repository ID
-        :param path: Absolute path of the Repostiory
-        """
-        self.logger.info('Searching for deps in repo')
-        self.logger.info(f'Repo ID: {repo_id}, Path: {path}')
-
-        deps = dep_calc.get_deps(path)
-
-        for dep in deps:
-                repo_deps = {
-                    'repo_id': repo_id,
-                    'dep_name' : dep.name,
-	                'dep_count' : dep.count,
-	                'dep_language' : dep.language,
-                    'tool_source': self.tool_source,
-                    'tool_version': self.tool_version,
-                    'data_source': self.data_source,
-                    'data_collection_date': datetime.now().strftime('%Y-%m-%dT%H:%M:%SZ')
-                }
-
-                result = self.db.execute(self.repo_dependencies_table.insert().values(repo_deps))
-                self.logger.info(f"Added dep: {result.inserted_primary_key}")
