+++ workers/github_worker/github_worker.py
@@ -67,7 +67,7 @@ class GitHubWorker(Worker):
             }
         }
 
-        source_issues = self.paginate_endpoint(
+        source_issues = self.new_paginate_endpoint(
             issues_url, action_map=action_map,
             table=self.issues_table, where_clause=self.issues_table.c.repo_id == self.repo_id
         )
@@ -196,7 +196,7 @@ class GitHubWorker(Worker):
         }
 
         # list to hold contributors needing insertion or update
-        issue_comments = self.paginate_endpoint(
+        issue_comments = self.new_paginate_endpoint(
             comments_url, action_map=comment_action_map, table=self.message_table,
             where_clause=self.message_table.c.msg_id.in_(
                 [
@@ -286,7 +286,7 @@ class GitHubWorker(Worker):
         self.logger.info(pd.DataFrame(pk_source_issues).columns)
         self.logger.info(pd.DataFrame(pk_source_issues))
         #list to hold contributors needing insertion or update
-        issue_events = self.paginate_endpoint(
+        issue_events = self.new_paginate_endpoint(
             events_url, table=self.issue_events_table, action_map=event_action_map,
             where_clause=self.issue_events_table.c.issue_id.in_(
                 set(pd.DataFrame(pk_source_issues)['issue_id'])
@@ -423,7 +423,7 @@ class GitHubWorker(Worker):
             }
         }
 
-        source_assignees_insert, _ = self.organize_needed_data(
+        source_assignees_insert, _ = self.new_organize_needed_data(
             assignees_all, augur_table=self.issue_assignees_table,
             action_map=assignee_action_map
         )
@@ -460,7 +460,7 @@ class GitHubWorker(Worker):
                 'augur': ['label_src_id']
             }
         }
-        source_labels_insert, _ = self.organize_needed_data(
+        source_labels_insert, _ = self.new_organize_needed_data(
             labels_all, augur_table=self.issue_labels_table,
             action_map=label_action_map
         )
