+++ workers/gitlab_issues_worker/gitlab_issues_worker.py
@@ -7,11 +7,8 @@ import sqlalchemy as s
 from workers.worker_base import Worker
 from urllib.parse import urlparse, quote
 
-#This needs to inherit from the git Interfaceable class
-from workers.worker_git_integration import WorkerGitInterfaceable
 
-
-class GitLabIssuesWorker(WorkerGitInterfaceable):
+class GitLabIssuesWorker(Worker):
     def __init__(self, config={}):
 
         worker_type = "gitlab_issues_worker"
