+++ workers/gitlab_merge_request_worker/gitlab_merge_request_worker.py
@@ -6,11 +6,8 @@ import sqlalchemy as s
 from workers.worker_base import Worker
 from urllib.parse import urlparse, quote
 
-#This class NEEDS to inherit from here.
-from workers.worker_git_integration import WorkerGitInterfaceable
 
-
-class GitlabMergeRequestWorker(WorkerGitInterfaceable):
+class GitlabMergeRequestWorker(Worker):
     def __init__(self, config={}):
         # Define what this worker can be given and know how to interpret
 
@@ -382,8 +379,6 @@ class GitlabMergeRequestWorker(WorkerGitInterfaceable):
                     AND LOWER(data_source) = '{} api'
                     """.format(pr_src_id, platform))
 
-        self.logger.info("DEV_DEBUG:" + str(pr_src_id))
-        self.logger.info("DEV_DEBUG:" + str(platform))
         self.logger.info(idSQL)
 
         rs = pd.read_sql(idSQL, self.db, params={})
