+++ workers/message_insights_worker/setup.py
@@ -28,7 +28,7 @@ setup(
         'requests==2.22.0',
         'psycopg2-binary==2.8.6',
         'click==7.1.2',
-        'scipy==1.7.0',
+        'scipy==1.4.1',
         'sklearn==0.0',
         'numpy==1.19.5',
         'nltk==3.5',
