+++ workers/release_worker/release_worker.py
@@ -9,11 +9,8 @@ from sqlalchemy import MetaData
 from sqlalchemy.ext.automap import automap_base
 from workers.worker_base import Worker
 
-#Breaks if doesn't inherit from GitInterfaceable
-from workers.worker_git_integration import WorkerGitInterfaceable
-
 #TODO - fully edit to match releases
-class ReleaseWorker(WorkerGitInterfaceable):
+class ReleaseWorker(Worker):
     """
     Worker that collects Repository Releases data from the Github API
     and stores it in our database.
